+++ next.config.mjs
@@ -12,6 +12,9 @@ const nextConfig = {
             },
         ];
     },
+    images: {
+        domains: ['ui-avatars.com'],
+    },
 };
 
 export default nextConfig;