+++ CHANGELOG.md
@@ -35,3 +35,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [t-28] Page profil utilisateur [#28](https://trello.com/c/JiNow5W7/28-t-28-page-profil-utilisateur)
 - [t-29] Page settings [#29](https://trello.com/c/qL72ZE9P/29-t-29-page-settings)
 - [t-30] profile settings : intégration des actions [#30](https://trello.com/c/efWXDAvv/30-t-30-profile-settings-int%C3%A9gration-des-actions)
+- [t-31] account settings : intégration des actions [#31](https://trello.com/c/XQ1DVZNn/31-t-31-account-settings-int%C3%A9gration-des-actions)